Check yarn version (#440)
* Add check for yarn version in createLessonPdfs, and require at least version 1.3.2

* Better error handling for createLessonPdfs.

* Process multiple PDFs simultaneously (currently 10 at a time).

+++ createLessonPdfs.js
@@ -1,19 +1,19 @@
 /* eslint-env node */
 
-const { spawn } = require('child_process');
+const { spawn, execSync } = require('child_process');
 const nodeCleanup = require('node-cleanup');
 const puppeteer = require('puppeteer');
 const path = require('path');
 const fse = require('fs-extra');
 const {buildDir, publicPath} = require('./buildconstants');
-
 const {lessonPaths} = require('./pathlists');
 
+const numberOfSimultaneousPdfConverts = 10;
 const urlBase = 'http://127.0.0.1:8080' + publicPath;
-
 let localWebServer = null;
+
 const cleanup = () => {
-  if (!localWebServer.killed) {
+  if (localWebServer && !localWebServer.killed) {
     console.log('Killing localWebServer');
     const killSignal = 'SIGTERM'; // 'SIGINT';
     const killSuccess = localWebServer.kill(killSignal);
@@ -23,57 +23,75 @@ const cleanup = () => {
     localWebServer.removeAllListeners();
   }
 };
-nodeCleanup(function (exitCode, signal) {
-  console.log('Exiting node script... (exitCode:' + exitCode + ', signal:' + signal + ')');
-  cleanup();
-});
 
 const convertUrl = async (browser, lesson) => {
-  try {
-    const pdfFile = path.join(buildDir, lesson + '.pdf');
-    const pdfFolder = path.dirname(pdfFile);
-    fse.mkdirsSync(pdfFolder);
-    const page = await browser.newPage();
-    const url = urlBase + lesson + '?pdf';
-    await page.goto(url, {waitUntil: 'networkidle'});
-    //await page.emulateMedia('screen');
-    console.log('Rendering PDF: ' + url + ' ---> ' + path.relative(__dirname, pdfFile));
-    await page.pdf({
-      path: pdfFile,
-      printBackground: true,
-      format: 'A4',
-      margin: {
-        top: '0.5in',
-        bottom: '0.5in',
-        left: '0.5in',
-        right: '0.5in',
-      }
-    });
-  } catch(err) {
-    console.log(err);
+  const pdfFile = path.join(buildDir, lesson + '.pdf');
+  const pdfFolder = path.dirname(pdfFile);
+  fse.mkdirsSync(pdfFolder);
+  const page = await browser.newPage();
+  const url = urlBase + lesson + '?pdf';
+  await page.goto(url, {waitUntil: 'networkidle'});
+  //await page.emulateMedia('screen');
+  console.log('Rendering PDF: ' + url + ' ---> ' + path.relative(__dirname, pdfFile));
+  await page.pdf({
+    path: pdfFile,
+    printBackground: true,
+    format: 'A4',
+    margin: {
+      top: '0.5in',
+      bottom: '0.5in',
+      left: '0.5in',
+      right: '0.5in',
+    }
+  });
+};
+
+const createChunks = (originalArray, chunkSize) => {
+  const chunks = [];
+  const arr = originalArray.slice();
+  while (arr.length > 0) {
+    chunks.push(arr.splice(0, chunkSize));
   }
+  return chunks;
 };
 
 const doConvert = () => {
   const lessons = lessonPaths();
-
   (async () => {
-    const browser = await puppeteer.launch();
+    try {
+      const browser = await puppeteer.launch();
+      const lessonChunks = createChunks(lessons, numberOfSimultaneousPdfConverts);
 
-    // Could perhaps look into Promise.race() to run several promises simultaneously
-    for (const path of lessons) {
-      await convertUrl(browser, path);
+      let pdfCount = 0;
+      const startTime = new Date().getTime();
+
+      for (const lessonChunk of lessonChunks) {
+        const convertPromises = lessonChunk.map(path => convertUrl(browser, path));
+        try {
+          await Promise.all(convertPromises); // Perhaps look into Promise.race() to run even faster
+          pdfCount += convertPromises.length;
+        } catch (err) {
+          console.log('Error while converting URLs. Skipping rest of lessons.');
+          console.log(err);
+          break;
+        }
+      }
+
+      const endTime = new Date().getTime();
+      const seconds = (endTime - startTime)/1000.0;
+      console.log('Time used to convert PDFs:', seconds, 'seconds = ', (seconds/pdfCount).toFixed(2), 'seconds/PDF');
+
+      browser.close();
     }
-
-    browser.close();
-    cleanup();
+    catch (e) {
+      console.log('Error in doConvert:', e);
+    }
+    finally {
+      cleanup();
+    }
   })();
 
 };
-
-localWebServer = spawn('yarn serve', {
-  shell: true
-});
 
 const checkStarted = (data) => {
   const str = String(data);
@@ -82,6 +100,31 @@ const checkStarted = (data) => {
     doConvert(localWebServer);
   }
 };
+
+const checkYarnVersion = () => {
+  const version = execSync('yarn --version', {shell: true}).toString().trim();
+  const lowestVersion = '1.3.2';
+  const [major = 0, minor = 0, patch = 0] = version.split('.');
+  const [lowestMajor = 0, lowestMinor = 0, lowestPatch = 0] = lowestVersion.split('.');
+  const tooLow = () => {
+    console.log('ERROR: The version of yarn (' + version + ') is too low. Must be >= ' + lowestVersion);
+    process.exit(1);
+  };
+  if (major < lowestMajor) { tooLow(); }
+  if (major === lowestMajor && minor < lowestMinor) { tooLow(); }
+  if (major === lowestMajor && minor === lowestMinor && patch < lowestPatch) { tooLow(); }
+};
+
+checkYarnVersion();
+
+nodeCleanup(function (exitCode, signal) {
+  console.log('Exiting node script... (exitCode:' + exitCode + ', signal:' + signal + ')');
+  cleanup();
+});
+
+localWebServer = spawn('yarn serve', {
+  shell: true
+});
 
 localWebServer.stdout.on('data', checkStarted);
 localWebServer.stderr.on('data', checkStarted);
